Cleaning and refactoring
