+++ package.json
@@ -44,11 +44,11 @@
     "testdouble-chai": "^0.5.0"
   },
   "devDependencies": {
-    "chai": "^6.2.2",
-    "chai-as-promised": "^8.0.2",    
     "@babel/core": "^7.28.6",
     "@babel/preset-env": "^7.28.6",
     "babel-jest": "^30.2.0",
+    "chai": "^6.2.2",
+    "chai-as-promised": "^8.0.2",
     "cross-env": "^10.1.0",
     "jest": "^30.2.0",
     "mocha": "^11.7.5",
